Feature/bookinghistory

+++ src/main/java/com/higgsup/base/dto/base/IPagedResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class IPagedResponse<T> {
   private ResponseMessage<T> responseMessage;
-  private int totalItem;
+  private long totalItem;
   private int pageIndex;
   private int pageSize;
 }
